+++ src/vcl-comps/Workspace.js
@@ -23,7 +23,10 @@ var Tab = {
         }
         
         this._nodes.text.innerHTML = html;
-    }
+    },
+	show: (tab) => tab.removeClass("tabs-hidden"),
+	hide: (tab) => !tab.hasClass("tabs-hidden") && tab.addClass("tabs-hidden"),
+	isHidden: (tab) => tab.hasClass("tabs-hidden")
 };
 var Utils = {
     getState: function(scope) {
@@ -88,6 +91,7 @@ var Utils = {
     }
 };
 
+
 $(["ui/Form"], {
     onLoad: function() {
         var scope = this.scope();
@@ -150,7 +154,7 @@ $(["ui/Form"], {
     	this.setSelected(false);
     }
 }, [
-    $(["devtools/TabFactory"], "editor-factory", {
+    $(["devtools/TabFactory"], ("editor-factory"), {
         vars: {
             closeable: true,
             formUri: "devtools/Editor<js>",
@@ -189,43 +193,44 @@ $(["ui/Form"], {
             return tab;
         }
     }),
-    $(("vcl/Action"), "editors-next", {
+    $("vcl/Action", ("editors-next"), {
     	onExecute: function() {
 			var tabs = this.up().qsa("#editors-tabs:visible");
-			// if(tabs.length === 1) {
-			// 	tabs = this._owner.qsa("Tabs<>:visible");
-			// }
 			var focused = this.up().vars("editors-tabs:focused");
 			if(focused && !focused.isFocused()) {
-				focused.app().print(focused, "fosuced");
-				return focused.setFocus();
-			}
-			if(tabs.length) {
+				// show(focused);
+				// focused.setFocus();
+			} else if(tabs.length) {
 				var index = tabs.indexOf(focused) + 1;
 				if(index >= tabs.length) index = 0;
 				focused = this.up().vars("editors-tabs:focused", tabs[index]);
-				focused.setFocus();
-				
-				focused.app().print(focused, "fosuced");
+				// if(Tab.isHidden(focused)) return;
+				// show(focused);
+				// focused.setFocus();
 			}
+			Tab.show(focused);
+			focused.setFocus();
     	}
     }),
-    $(("vcl/Action"), "editors-previous", {
+    $("vcl/Action", ("editors-previous"), {
     	onExecute: function() {
 			var tabs = this.up().qsa("#editors-tabs:visible");
 			var focused = this.up().vars("editors-tabs:focused");
 			if(focused && !focused.isFocused()) {
-				return focused.setFocus();
-			}
-			if(tabs.length) {
+				// show(focused);
+				// focused.setFocus();
+			} else if(tabs.length) {
 				var index = tabs.indexOf(focused) - 1;
 				if(index < 0) index = tabs.length - 1;
 				focused = this.up().vars("editors-tabs:focused", tabs[index]);
-				focused.setFocus();
+				// show(focused);
+				// focused.setFocus();
 			}
+			Tab.show(focused);
+			focused.setFocus();
     	}
     }),
-    $(("vcl/Action"), "editors-close-all", {
+    $("vcl/Action", ("editors-close-all"), {
     	onExecute: function(evt) {
 			var tabs = this.vars(["editors-tabs:focused"]) || this.udown("#editors-tabs");
             var selected = tabs.getSelectedControl(1);
@@ -241,14 +246,14 @@ $(["ui/Form"], {
 	    		});
     	}
     }),
-    $(("vcl/Action"), "editor-new", {
+    $("vcl/Action", ("editor-new"), {
         onExecute: function(evt) {
             this.scope("editor-factory")
             	.execute(evt)
             	.setSelected(true);
         }
     }),
-    $(("vcl/Action"), "editor-needed", {
+    $("vcl/Action", ("editor-needed"), {
         onExecute: function(evt) {
             var scope = this.scope(), tab;
             if(typeof evt === "string") {
@@ -307,7 +312,7 @@ $(["ui/Form"], {
     		return tab;
         }
     }),
-    $(("vcl/Action"), "editor-next", {
+    $("vcl/Action", ("editor-next"), {
         onExecute: function() {
         	var ws = this.up();
 			var tabs = ws.vars("editors-tabs:focused") || ws.qs("#editors-tabs");
@@ -318,7 +323,7 @@ $(["ui/Form"], {
 			tabs.selectNext();
         }
     }),
-    $(("vcl/Action"), "editor-previous", {
+    $("vcl/Action", ("editor-previous"), {
         onExecute: function() {
         	var ws = this.up();
 			var tabs = ws.vars("editors-tabs:focused") || ws.qs("#editors-tabs");
@@ -329,7 +334,7 @@ $(["ui/Form"], {
 			tabs.selectPrevious();
         }
     }),
-    $(("vcl/Action"), "editor-close", {
+    $("vcl/Action", ("editor-close"), {
         onExecute: function(evt) {
             var scope = this.getScope();
             var selected = scope['editors-tabs'].getSelectedControl(1);
@@ -339,12 +344,12 @@ $(["ui/Form"], {
             evt.preventDefault();
         }
     }),
-    $(("vcl/Action"), "editor-move-to-front", {
+    $("vcl/Action", ("editor-move-to-front"), {
     	onExecute: function() {
     		this._owner.qs("vcl/ui/Tab:selected:childOf(editors-tabs)").setIndex(0);
     	}
     }),
-    $(("vcl/Action"), "editor-move-left", {
+    $("vcl/Action", ("editor-move-left"), {
     	onExecute: function() {
     		var tab = this._owner.qs("vcl/ui/Tab:selected:childOf(editors-tabs)");
     		var index = tab.getIndex();
@@ -353,19 +358,19 @@ $(["ui/Form"], {
     		}
     	}
     }),
-    $(("vcl/Action"), "editor-move-right", {
+    $("vcl/Action", ("editor-move-right"), {
     	onExecute: function() {
     		var tab = this._owner.qs("vcl/ui/Tab:selected:childOf(editors-tabs)");
     		var index = tab.getIndex();
     		tab.setIndex(index + 1);
     	}
     }),
-    $(("vcl/Action"), "editor-setfocus", {
+    $("vcl/Action", ("editor-setfocus"), {
     	onExecute: function(evt) {
 			this._owner.qs("vcl/ui/Tab:selected:childOf(editors-tabs) #ace").setFocus();
     	}
     }),
-    $(("vcl/Action"), "editor-focus-in-navigator", {
+    $("vcl/Action", ("editor-focus-in-navigator"), {
         onExecute: function(evt) {
         	// TODO 
             // var app = this.getApp();
@@ -375,7 +380,7 @@ $(["ui/Form"], {
         }
     }),
     
-    $("vcl/ui/Panel#left-sidebar", { align: "left", css: "border-right: 1px solid gray;", width: 375 }, [
+    $("vcl/ui/Panel", ("left-sidebar"), { align: "left", css: "border-right: 1px solid gray;", width: 375 }, [
     	
         $("vcl/ui/Tabs#left-sidebar-tabs", [
             $("vcl/ui/Tab", { text: locale("Navigator"), control: "navigator", selected: true }),
@@ -399,8 +404,7 @@ $(["ui/Form"], {
         $("vcl/ui/Panel", "search-panel", { align: "client", visible: false }),
         $("vcl/ui/Panel", "inspector-panel", { align: "client", visible: false })
     ]),
-
-    $("vcl/ui/Panel#editors", { align: "client", css: "background-color: silver;" }, [
+    $("vcl/ui/Panel", ("editors"), { align: "client", css: "background-color: silver;" }, [
         $("vcl/ui/Tabs", "editors-tabs", {
             onChange: function(tab, previous) {
 // TODO tell application to render it's title